requirements.txt
appnope==0.1.0
ase==3.19.1
autopep8==1.5
backcall==0.1.0
boto3==1.12.23
botocore==1.15.23
Brotli==1.0.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
ctgan==0.2.1
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
Flask==1.1.2
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.3
future==0.18.2
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googledrivedownloader==0.4
h5py==2.10.0
idna==2.9
imageio==2.8.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
jupyter-client==6.0.0
jupyter-core==4.6.3
karateclub==0.45.27
kiwisolver==1.1.0
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.1.3
networkx==2.4
numba==0.48.0
numpy==1.18.1
ogb==1.1.1
pandas==0.25.3
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.8.1
plyfile==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.5.2
PyGSP==0.5.1
PyMySQL==0.9.3
pyparsing==2.4.6
python-dateutil==2.8.1
python-louvain==0.13
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
rdflib==4.2.2
requests==2.23.0
retrying==1.3.3
rsa==4.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sklearn-pandas==1.8.0
smart-open==1.10.0
SQLAlchemy==1.3.17
torch==1.5.0
torch-geometric==1.5.0
torch-scatter==2.0.4
torch-sparse==0.6.5
torchvision==0.5.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.1
